+++ apps/docs/sidebars.ts
@@ -31,13 +31,6 @@ const sidebars: SidebarsConfig = {
     },
     {
       type: "category",
-      label: "Guides",
-      items: [
-        "guides/projects-on-org-scope",
-      ],
-    },
-    {
-      type: "category",
       label: "Block Sequencing",
       items: [
         "sequencers/overview",
@@ -111,6 +104,16 @@ const sidebars: SidebarsConfig = {
       items: [
         "client/overview",
         "client/react",
+      ],
+    },
+    {
+      type: "category",
+      label: "UI Components",
+      items: [
+        "ui/overview",
+        "ui/common-components",
+        "ui/flow-aware-components",
+        "ui/generative",
       ],
     },
     {